several test fixes
- renamed test classes so the type tested is in the prefix only
- GLTF APK tests weren't being compiled due to .asmdef misconfiguration
- tests were failing because test scene wasn't being loaded
