+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/AssetManager/AssetBundles/AB_GameObject/Tests/APK_AB_GameObject_ShouldWorkWhen.cs
@@ -4,7 +4,7 @@ using DCL.Helpers;
 
 namespace AssetPromiseKeeper_AssetBundle_GameObject_Tests
 {
-    public class APKShouldWorkWhen_AB_GameObject :
+    public class APK_AB_GameObject_ShouldWorkWhen :
         APKWithPoolableAssetShouldWorkWhen_Base<AssetPromiseKeeper_AB_GameObject,
                                                 AssetPromise_AB_GameObject,
                                                 Asset_AB_GameObject,
