+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/AssetManager/Common/Mock/Tests/AssetPromiseKeeperTests.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "AssetPromiseKeeperTests",
+    "name": "AssetPromiseKeeper_Basic_Tests",
     "references": [
         "MainScripts",
         "UnityGLTFAssembly",
