+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/AssetManager/Common/Tests/TestsBase_APK.cs
@@ -15,14 +15,14 @@ namespace AssetPromiseKeeper_Tests
         protected APKType keeper;
 
         [UnitySetUp]
-        protected IEnumerator SetUp()
+        protected virtual IEnumerator SetUp()
         {
             keeper = new APKType();
             yield break;
         }
 
         [UnityTearDown]
-        protected IEnumerator TearDown()
+        protected virtual IEnumerator TearDown()
         {
             PoolManager.i.Cleanup();
             keeper.Cleanup();
