+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/AssetManager/GLTF/Tests/APK_GLTF_Tests.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "APK_GLTF_Tests",
+    "name": "AssetPromiseKeeper_GLTF_Tests",
     "references": [
         "GUID:e555144732b726c4cba5b0f6337fea75",
         "GUID:b24824f3bc2a60641a01e2083614f802",
@@ -10,9 +10,7 @@
     "optionalUnityReferences": [
         "TestAssemblies"
     ],
-    "includePlatforms": [
-        "Editor"
-    ],
+    "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
     "overrideReferences": false,
